requirements.txt
aiohttp
discord.py
pydantic
pydantic-settings
diffusers>=0.29.0
transformers[sentencepiece]>=4.41.0,<5.0.0
accelerate>=0.30.0
torch
numpy
Pillow
sentencepiece>=0.2.0
protobuf>=3.20.0
PyNaCl
bitsandbytes>=0.43.0
